Add account context

+++ db/accountService.js
@@ -2,7 +2,7 @@ import { initDB } from "./database";
 
 const getAccounts = async (userId) => {
   try {
-    const db = initDB();
+    const db = await initDB();
     const result = await db.getAllAsync(
       "SELECT * FROM accounts WHERE user_id = ?",
       [userId],
@@ -10,14 +10,28 @@ const getAccounts = async (userId) => {
 
     return result;
   } catch (error) {
-    console.error("Failed to fetch accounts. ERR:", error);
+    console.error("[Database] Failed to fetch accounts. ERR: ", error);
     throw error;
   }
 };
 
-const addAccount = async (account) => {
+const getSingleAccount = async (accountId) => {
+  try {
+    const db = await initDB();
+    const result = await db.getFirstAsync(
+      "SELECT * FROM accounts WHERE id = ?",
+      [accountId],
+    );
+
+    return result;
+  } catch (error) {
+    console.error("[Database] Failed to fetch account. ERR: ", error);
+    throw error;
+  }
+};
+
+const addAccount = async (account, userId) => {
   const {
-    user_id,
     name,
     currency,
     type,
@@ -28,11 +42,11 @@ const addAccount = async (account) => {
   } = account;
 
   try {
-    const db = initDB();
+    const db = await initDB();
     const result = await db.runAsync(
-      "INSERT INTO accounts (user_id, name, currency, type, amount, credit) VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
+      "INSERT INTO accounts (user_id, name, currency, type, amount, total_income, total_expense, credit) VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
       [
-        user_id,
+        userId,
         name,
         currency,
         type,
@@ -43,87 +57,75 @@ const addAccount = async (account) => {
       ],
     );
 
-    return result;
+    return { id: result.lastInsertRowId };
   } catch (error) {
-    console.error("Failed to add account. ERR: ", error);
+    console.error("[Database] Failed to add account. ERR: ", error);
     throw error;
   }
 };
 
 const editAccount = async (account) => {
-  const { user_id, id, name, currency, type, credit } = account;
+  const { id, name, currency, type, credit } = account;
 
   try {
-    const db = initDB();
-    const result = await db.runAsync(
-      "UPDATE accounts SET name = ?, currency = ?, type = ?, credit = ? WHERE user_id = ? AND id = ?",
-      [name, currency, type, credit, user_id, id],
+    const db = await initDB();
+    await db.runAsync(
+      "UPDATE accounts SET name = ?, currency = ?, type = ?, credit = ? WHERE id = ?",
+      [name, currency, type, credit, id],
     );
-
-    return result;
   } catch (error) {
-    console.error("Failed to edit account. ERR: ", error);
+    console.error("[Database] Failed to edit account. ERR: ", error);
     throw error;
   }
 };
 
 const updateAccountAmount = async (accountId, transactionAmount, type) => {
   try {
-    const db = initDB();
+    const db = await initDB();
     const accountInfo = await db.getFirstAsync(
       "SELECT amount, total_income, total expense FROM accounts WHERE id = ?",
       [accountId],
     );
 
+    let newAmount = accountInfo.amount;
+    let newIncome = accountInfo.total_income;
+    let newExpense = accountInfo.total_expense;
+
     if (type === "Income") {
-      const newAmount = accountInfo.amount + transactionAmount;
-      const newIncome = accountInfo.total_income + transactionAmount;
-
-      const result = await db.runAsync(
-        "UPDATE accounts SET amount = ?, total_income = ? WHERE id = ?",
-        [newAmount, newIncome, accountId],
-      );
-
-      return result;
+      newAmount += transactionAmount;
+      newIncome += transactionAmount;
     } else {
-      const newAmount = accountInfo.amount - transactionAmount;
-      const newExpense = accountInfo.total_expense + transactionAmount;
-
-      const result = await db.runAsync(
-        "UPDATE accounts SET amount = ?, total_expense =? WHERE id = ?",
-        [newAmount, newExpense, accountId],
-      );
-
-      return result;
+      newAmount -= transactionAmount;
+      newExpense += transactionAmount;
     }
+
+    await db.runAsync(
+      "UPDATE accounts SET amount = ?, total_income = ?, total_expense = ? WHERE id = ?",
+      [newAmount, newIncome, newExpense, accountId],
+    );
   } catch (error) {
-    console.error("Failed to update account amount. ERR: ", error);
+    console.error("[Database] Failed to update account amount. ERR: ", error);
     throw error;
   }
 };
 
 const deleteAccount = async (accountId) => {
   try {
-    const db = initDB();
-    const resultTransactions = await db.runAsync(
-      "DELETE FROM transactions WHERE account_id = ?",
-      [accountId],
-    );
+    const db = await initDB();
+    await db.runAsync("DELETE FROM transactions WHERE account_id = ?", [
+      accountId,
+    ]);
 
-    const resultAccounts = await db.runAsync(
-      "DELETE FROM accounts WHERE id = ?",
-      [accountId],
-    );
-
-    return { resultTransactions, resultAccounts };
+    await db.runAsync("DELETE FROM accounts WHERE id = ?", [accountId]);
   } catch (error) {
-    console.error("Failed to delete account. ERR: ", error);
+    console.error("[Database] Failed to delete account. ERR: ", error);
     throw error;
   }
 };
 
 export {
   getAccounts,
+  getSingleAccount,
   addAccount,
   editAccount,
   updateAccountAmount,
